requirements.txt
langchain
huggingface_hub
supabase
python-dotenv
langchain-community
google-generativeai
pandas
pillow
